+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 - More advancement helper stuff
 - BaseScreen#buttonClicked
 - ClientClickedMenuButton#isShiftDown
-- <T : Enum<T>>.nextEnum()
-- <T : Enum<T>>.prevEnum()
+- <T : Enum<T>>#nextEnum()
+- <T : Enum<T>>#prevEnum()
+- DyeColor#getNameComponent()
 
 # 1.19.0